backend/app/requirements.txt
fastapi==0.99.1
mangum==0.17.0
pymysql==1.1.1
passlib[bcrypt]
python-multipart==0.0.9
python-dotenv==0.19.1
bcrypt==4.1.3
boto3
pydantic[email]